+++ dbRangley/functions/rgl_fn_v_user_profile_by_uuid.sql
@@ -4,18 +4,18 @@ CREATE OR REPLACE FUNCTION rangley.rgl_fn_v_user_profile_by_uuid
 )
 RETURNS table
 (
-    user_uuid               uuid,
-    username                varchar(50),
-    display_name            varchar(50),
-    member_since            timestamptz,
-    meets_created           bigint,
-    meets_attended          bigint,
-    friend_count            bigint,
-    discoverable_by_username boolean,
-    discoverable_by_phone   boolean,
-    discoverable_by_email   boolean,
-    show_full_name          boolean,
-    allow_invites_from_anyone boolean
+    user_uuid               	uuid,
+    username                	varchar(50),
+    display_name            	varchar(50),
+    member_since            	timestamptz,
+    meets_created           	INT8,
+    meets_attended          	INT8,
+    friend_count            	INT8,
+    discoverable_by_username 	boolean,
+    discoverable_by_phone   	boolean,
+    discoverable_by_email   	boolean,
+    show_full_name          	boolean,
+    allow_invites_from_anyone 	boolean
 )
 LANGUAGE sql
 STABLE
